Fixed project files after the directory changes. Only major bug in the solution currently is that earlier projects use a different Input Layout that do not have a Tangent component of the Vertex structure. So geometry objects, like GCube and GPlane, will have errors. You must comment out the related Tangent line in these geometry objects in order to run early demos.

+++ Common/Utility/GFirstPersonCamera.cpp
@@ -90,7 +90,7 @@ float GFirstPersonCamera::GetFovY()const
 float GFirstPersonCamera::GetFovX()const
 {
 	float halfWidth = 0.5f*GetNearWindowWidth();
-	return 2.0f*atan(halfWidth / mNearZ);
+	return 2.0f*atanf(halfWidth / mNearZ);
 }
 
 float GFirstPersonCamera::GetNearWindowWidth()const
